chore: add keytype for pkcs12:export()
`keytype` adds a flag to the store private key. This is a non standard extension
that is only currently interpreted by MSIE. If set to zero the flag is omitted,
if set to `KEY_SIG` the key can be used for signing only, if set to `KEY_EX`
it can be used for signing and encryption. This option was useful for old
export grade software which could use signing only keys of arbitrary size but
had restrictions on the permissible sizes of keys which could be used for
encryption.